Add control_host_iot/_kvm connectors as copies of zapper_*
Add control_host_iot and control_host_kvm as verbatim copies of the
zapper_iot and zapper_kvm connectors, and register them in
DEVICE_CONNECTORS. The old zapper_iot/zapper_kvm modules become thin
shims that re-export DeviceConnector from the new modules and emit a
DeprecationWarning, so existing agent configs keep working.
