+++ device-connectors/src/testflinger_device_connectors/devices/__init__.py
@@ -53,6 +53,8 @@ DEVICE_CONNECTORS = (
     "noprovision",
     "oemrecovery",
     "oemscript",
+    "control_host_iot",
+    "control_host_kvm",
     "zapper_iot",
     "zapper_kvm",
 )
