+++ device-connectors/src/testflinger_device_connectors/devices/control_host_kvm/__init__.py
@@ -1,0 +1,278 @@
+# Copyright (C) 2026 Canonical
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Zapper Connector for KVM provisioning."""
+
+import base64
+import binascii
+import contextlib
+import logging
+import subprocess
+from pathlib import Path
+from typing import Any, Dict, Optional, Tuple
+
+import yaml
+
+from testflinger_device_connectors.devices import ProvisioningError
+from testflinger_device_connectors.devices.dell_oemscript import DellOemScript
+from testflinger_device_connectors.devices.hp_oemscript import HPOemScript
+from testflinger_device_connectors.devices.lenovo_oemscript import (
+    LenovoOemScript,
+)
+from testflinger_device_connectors.devices.oemscript import OemScript
+from testflinger_device_connectors.devices.zapper import ZapperConnector
+
+logger = logging.getLogger(__name__)
+
+JAMMY_OEM_PRESET = "desktop-jammy-oem"
+
+
+class DeviceConnector(ZapperConnector):
+    """Tool for provisioning baremetal with a given image."""
+
+    PROVISION_METHOD = "ProvisioningKVM"
+
+    def _validate_base_user_data(self, encoded_user_data: str):
+        """Assert `base_user_data` argument is a valid base64 encoded YAML."""
+        try:
+            user_data = base64.b64decode(encoded_user_data.encode()).decode()
+            yaml.safe_load(user_data)
+        except (binascii.Error, ValueError) as exc:
+            raise ProvisioningError(
+                "Provided `base_user_data` is not base64 encoded."
+            ) from exc
+        except yaml.YAMLError as exc:
+            raise ProvisioningError(
+                "Provided `base_user_data` is not a valid YAML."
+            ) from exc
+
+    def _get_autoinstall_conf(self) -> Optional[Dict[str, Any]]:
+        """Autoinstall-related keys are pre-fixed with `autoinstall_`.
+
+        If any of those arguments are provided and valid, the function
+        returns an autoinstall_conf dictionary.
+        """
+        autoinstall_conf = {}
+
+        if self.job_data["provision_data"].get("autoinstall_oem"):
+            # In case of Ubuntu OEM via autoinstall, use a
+            # dedicated user-data file as base.
+
+            logger.info(
+                "When using 'autoinstall_oem', "
+                "other autoinstall keys are not considered"
+            )
+
+            user_data_oem = (
+                Path(__file__).parent / "../../data/zapper_kvm/user-data-oem"
+            )
+            user_data = user_data_oem.read_text(encoding="utf-8")
+            encoded_user_data = base64.b64encode(user_data.encode()).decode()
+
+            user_data = yaml.safe_load(user_data)
+            storage = user_data["autoinstall"]["storage"]["layout"]["name"]
+
+            autoinstall_conf = {
+                "base_user_data": encoded_user_data,
+                "storage_layout": storage,
+            }
+
+        else:
+            for key, value in self.job_data["provision_data"].items():
+                if "autoinstall_" not in key:
+                    continue
+
+                autoinstall_key = key.replace("autoinstall_", "")
+                with contextlib.suppress(AttributeError):
+                    getattr(self, f"_validate_{autoinstall_key}")(value)
+                autoinstall_conf[autoinstall_key] = value
+
+        if not autoinstall_conf:
+            logger.info("Autoinstall-related keys were not provided.")
+            return None
+
+        return autoinstall_conf
+
+    def _read_ssh_key(self) -> str:
+        return (
+            Path("~/.ssh/id_rsa.pub").expanduser().read_text(encoding="utf-8")
+        )
+
+    def _get_credentials(self, target: str | None = None) -> tuple[str, str]:
+        if target == JAMMY_OEM_PRESET:
+            return "ubuntu", "u"
+        else:
+            return (
+                self.job_data.get("test_data", {}).get(
+                    "test_username", "ubuntu"
+                ),
+                self.job_data.get("test_data", {}).get(
+                    "test_password", "ubuntu"
+                ),
+            )
+
+    OPTIONAL_KEYS = (
+        "cmdline_append",
+        "skip_download",
+        "wait_until_ssh",
+        "live_image",
+        "ubuntu_sso_email",
+        "robot_retries",
+    )
+
+    def _validate_configuration(
+        self,
+    ) -> Tuple[Tuple, Dict[str, Any]]:
+        """Validate the job config and data and prepare the arguments
+        for the Zapper `provision` API.
+        """
+        job_provision = self.job_data["provision_data"]
+
+        if "preset" in job_provision:
+            provision_data = self._build_preset_payload()
+        elif "alloem_url" in job_provision:
+            provision_data = self._build_alloem_payload()
+        else:
+            provision_data = self._build_default_payload()
+
+        provision_data["authorized_keys"] = [self._read_ssh_key()]
+        # Let Zapper handle defaults — only forward explicitly-set keys.
+        provision_data.update(
+            {
+                opt: job_provision[opt]
+                for opt in self.OPTIONAL_KEYS
+                if opt in job_provision
+            }
+        )
+
+        return ((), provision_data)
+
+    def _build_preset_payload(self) -> Dict[str, Any]:
+        job_provision = self.job_data["provision_data"]
+        preset = job_provision["preset"]
+
+        provision_data = {
+            key: value
+            for key, value in job_provision.items()
+            if not key.startswith("autoinstall")
+        }
+        has_autoinstall = any(
+            key.startswith("autoinstall") for key in job_provision
+        )
+
+        provision_data["username"], provision_data["password"] = (
+            self._get_credentials(preset)
+        )
+        if has_autoinstall:
+            provision_data["autoinstall_conf"] = self._get_autoinstall_conf()
+
+        if preset == JAMMY_OEM_PRESET:
+            # control host only does recovery
+            provision_data.pop("url", None)
+            if alloem_url := job_provision.get("alloem_url"):
+                provision_data["url"] = alloem_url
+                provision_data.pop("alloem_url", None)
+
+        return provision_data
+
+    def _build_alloem_payload(self) -> Dict[str, Any]:
+        job_provision = self.job_data["provision_data"]
+        username, password = self._get_credentials(JAMMY_OEM_PRESET)
+        return {
+            "url": job_provision["alloem_url"],
+            "username": username,
+            "password": password,
+            "autoinstall_conf": self._get_autoinstall_conf(),
+            "robot_tasks": job_provision["robot_tasks"],
+        }
+
+    def _build_default_payload(self) -> Dict[str, Any]:
+        job_provision = self.job_data["provision_data"]
+        username, password = self._get_credentials()
+        return {
+            "url": job_provision["url"],
+            "username": username,
+            "password": password,
+            "autoinstall_conf": self._get_autoinstall_conf(),
+            "robot_tasks": job_provision["robot_tasks"],
+        }
+
+    def _post_run_actions(self, args):
+        super()._post_run_actions(args)
+
+        provision_data = self.job_data["provision_data"]
+        used_alloem = "alloem_url" in provision_data
+        is_jammy_oem_preset = provision_data.get("preset") == JAMMY_OEM_PRESET
+
+        if used_alloem or is_jammy_oem_preset:
+            self._post_run_actions_oem(args)
+
+    def _post_run_actions_oem(self, args):
+        """Post run actions for 22.04 OEM images."""
+        try:
+            self._change_password("ubuntu", "u")
+            self._copy_ssh_id()
+        except subprocess.CalledProcessError as exc:
+            logger.error("Process failed with: %s", exc.output.decode())
+            raise ProvisioningError(
+                "Failed configuring SSH on the DUT."
+            ) from exc
+        except subprocess.TimeoutExpired as exc:
+            raise ProvisioningError(
+                "Timed out configuring SSH on the DUT."
+            ) from exc
+
+        self._run_oem_script(args)
+
+    def _run_oem_script(self, args):
+        """If "alloem_url" was in scope, the Zapper only restored
+        the OEM reset partition. The usual oemscript will take care
+        of the rest.
+        """
+        if not self.job_data["provision_data"].get("url"):
+            logger.warning(
+                "Provisioned with base `alloem` image, no test URL specified."
+            )
+            return
+
+        oem = self.job_data["provision_data"].get("oem")
+        oemscript = {
+            "hp": HPOemScript,
+            "dell": DellOemScript,
+            "lenovo": LenovoOemScript,
+        }.get(oem, OemScript)(args.config, args.job_data)
+
+        oemscript.provision()
+
+    def _change_password(self, username, orig_password):
+        """Change password via SSH to the one specified in the job data."""
+        password = self.job_data.get("test_data", {}).get(
+            "test_password", "ubuntu"
+        )
+        logger.info("Changing the original password to %s", password)
+
+        cmd = [
+            "sshpass",
+            "-p",
+            orig_password,
+            "ssh",
+            "-o",
+            "StrictHostKeyChecking=no",
+            "-o",
+            "UserKnownHostsFile=/dev/null",
+            f"{username}@{self.config['device_ip']}",
+            f"echo 'ubuntu:{password}' | sudo chpasswd",
+        ]
+
+        subprocess.check_output(cmd, stderr=subprocess.STDOUT, timeout=60)
